requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.6
django-debug-toolbar==5.0.1
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.9
et_xmlfile==2.0.0
inflection==0.5.1
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
